Configuration of Model Diagnostics

Diagnostics_cfg is a configuration class for estimating a variety of diagnostics for the models trained in the course of HTE estimation.

Public fields

ps

Model diagnostics for the propensity score model.

outcome

Model diagnostics for the outcome models.

effect

Model diagnostics for the joint effect model.

params

Parameters for any requested diagnostics.

Methods

Public methods

Diagnostics_cfg$new()

Diagnostics_cfg$add()

Diagnostics_cfg$clone()

Method new()

Create a new Diagnostics_cfg object with specified diagnostics to estimate.

Usage

Diagnostics_cfg$new(ps = NULL, outcome = NULL, effect = NULL, params = NULL)

Arguments

ps

Model diagnostics for the propensity score model.

outcome

Model diagnostics for the outcome models.

effect

Model diagnostics for the joint effect model.

params

List providing values for parameters to any requested diagnostics.

Returns

A new Diagnostics_cfg object.

Examples

Diagnostics_cfg$new(
   outcome = c("SL_risk", "SL_coefs", "MSE", "RROC"),
   ps = c("SL_risk", "SL_coefs", "AUC")
)

Method add()

Add diagnostics to the Diagnostics_cfg object.

Usage

Diagnostics_cfg$add(ps = NULL, outcome = NULL, effect = NULL)

Arguments

ps

Model diagnostics for the propensity score model.

outcome

Model diagnostics for the outcome models.

effect

Model diagnostics for the joint effect model.

Returns

An updated Diagnostics_cfg object.

Examples

cfg <- Diagnostics_cfg$new(
   outcome = c("SL_risk", "SL_coefs", "MSE", "RROC"),
   ps = c("SL_risk", "SL_coefs")
)
cfg <- cfg$add(ps = "AUC")

Method clone()

The objects of this class are cloneable with this method.

Usage

Diagnostics_cfg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

Diagnostics_cfg$new(
   outcome = c("SL_risk", "SL_coefs", "MSE", "RROC"),
   ps = c("SL_risk", "SL_coefs", "AUC")
)
#> <Diagnostics_cfg>
#>   Public:
#>     add: function (ps = NULL, outcome = NULL, effect = NULL) 
#>     clone: function (deep = FALSE) 
#>     effect: 
#>     initialize: function (ps = NULL, outcome = NULL, effect = NULL, params = NULL) 
#>     outcome: SL_risk SL_coefs MSE RROC
#>     params: list
#>     ps: SL_risk SL_coefs AUC

## ------------------------------------------------
## Method `Diagnostics_cfg$new`
## ------------------------------------------------

Diagnostics_cfg$new(
   outcome = c("SL_risk", "SL_coefs", "MSE", "RROC"),
   ps = c("SL_risk", "SL_coefs", "AUC")
)
#> <Diagnostics_cfg>
#>   Public:
#>     add: function (ps = NULL, outcome = NULL, effect = NULL) 
#>     clone: function (deep = FALSE) 
#>     effect: 
#>     initialize: function (ps = NULL, outcome = NULL, effect = NULL, params = NULL) 
#>     outcome: SL_risk SL_coefs MSE RROC
#>     params: list
#>     ps: SL_risk SL_coefs AUC

## ------------------------------------------------
## Method `Diagnostics_cfg$add`
## ------------------------------------------------

cfg <- Diagnostics_cfg$new(
   outcome = c("SL_risk", "SL_coefs", "MSE", "RROC"),
   ps = c("SL_risk", "SL_coefs")
)
cfg <- cfg$add(ps = "AUC")