Calculate Variable Importance of HTEs

calculate_vimp estimates the reduction in (population) $R^2$ from removing a particular moderator from a model containing all moderators.

calculate_vimp(.data, weight_col, pseudo_outcome, ..., .VIMP_cfg, .Model_cfg)

Arguments

.data

dataframe

weight_col

Unquoted name of the weight column.

pseudo_outcome

Unquoted name of the pseudo-outcome.

...

Unquoted names of covariates to include in the joint effect model. The variable importance will be calculated for each of these covariates.

.VIMP_cfg

A VIMP_cfg object defining how VIMP should be estimated.

.Model_cfg

A Model_cfg object defining how the joint effect model should be estimated.

References

Williamson, BD, Gilbert, PB, Carone, M, Simon, N. Nonparametric variable importance assessment using machine learning techniques. Biometrics. 2021; 77: 9-- 22. doi: 10.1111/biom.13392