Construct Pseudo-outcomes

construct_pseudo_outcomes takes a dataset which has been prepared with plugin estimators of nuisance parameters and transforms these into a "pseudo-outcome": an unbiased estimator of the conditional average treatment effect under exogeneity.

construct_pseudo_outcomes(.data, outcome, treatment, type = "dr")

Arguments

.data

dataframe (already prepared with attach_config, make_splits, and produce_plugin_estimates)

outcome

Unquoted name of outcome variable.

treatment

Unquoted name of treatment variable.

type

String representing how to construct the pseudo-outcome. Valid values are "dr" (the default), "ipw" and "plugin". See "Details" for more discussion of these options.