Function to calculate diagnostics based on model outputs

This function defines the calculations of common model diagnostics which are available.

estimate_diagnostic(.data, label, prediction, diag_name, params)

Arguments

.data

The full data frame with all auxilliary columns.

label

The (string) column name for the labels to evaluate against.

prediction

The (string) column name of predictions from the model to diagnose.

diag_name

The (string) name of the diagnostic to calculate. Currently available are "AUC", "MSE", "SL_coefs", "SL_risk", "RROC"

params

Any other necessary options to pass to the given diagnostic.

Examples

df <- dplyr::tibble(y = rbinom(100, 1, 0.5), p = rep(0.5, 100))
estimate_diagnostic(df, "y", "p", "AUC")
#> Error in estimate_diagnostic(df, "y", "p", "AUC"): could not find function "estimate_diagnostic"