Fits a T-learner using the appropriate settings

This function prepares data, fits the appropriate model and returns the resulting estimates in a standardized format.

fit_plugin_Y(full_data, weight_col, y_col, a_col, ..., .Model_cfg)

Arguments

full_data

The full dataset of interest for the modelling problem.

weight_col

The unquoted weighting variable name to use in model fitting.

y_col

The unquoted column name of the outcome.

a_col

The unquoted column name of the treatment.

...

The unquoted names of covariates to use in the model.

.Model_cfg

A Model_cfg object configuring the appropriate model type to use.

Value

A list with two elements, mu1 and mu0 corresponding to the models fit to the treatment and control potential outcomes, respectively. Each is a new Predictor object of the appropriate subclass corresponding to the the Model_cfg fit to the data.